requirements.txt
# Requirements for Tranquil-Inbox-Ward
# Generated from imports in app.py
python-dotenv
fastapi
uvicorn
requests
pydantic
